Start Removal Of PyTorch-only Code
